refactor(sdk/analyze): extract shell tokenizer to patterns/shell.rs
patterns.rs carried a ~250-line self-contained shell-command parser (segment
splitting, quote-aware tokenization, redirect/operand detection) that only the
edit-heavy / codex-read detectors use through one entry point. Move it to a
patterns/shell.rs submodule exposing just shell_command_has_file_read; the
codex read-command vocabulary (is_codex_shell_read_command) moves with it since
nothing else references it. patterns.rs drops from 1851 to 1578 lines.

Pure code move (byte-identical apart from the one pub(super) on the entry fn);
992/992 workspace tests pass.
